+++ CopilotKit/examples/next-openai/src/app/page.tsx
@@ -1,12 +1,16 @@
 "use client";
 
-import { CopilotProvider } from "@copilotkit/react-core";
+import { CopilotProvider, CopilotApiConfig } from "@copilotkit/react-core";
 import { CopilotSidebarUIProvider } from "@copilotkit/react-ui";
 import { VacationList } from "./components/vacation-list";
 
 export default function WaterBnb() {
   return (
-    <CopilotProvider>
+    <CopilotProvider
+      copilotApiConfig={{
+        chatApiEndpoint: "/api/copilotkit/chat",
+      }}
+    >
       <CopilotSidebarUIProvider>
         <VacationList />
       </CopilotSidebarUIProvider>