+++ CopilotKit/packages/react-core/src/hooks/index.ts
@@ -4,3 +4,4 @@ export type { UseCopilotChatReturn } from "./use-copilot-chat";
 
 export { useMakeCopilotActionable } from "./use-make-copilot-actionable";
 export { useMakeCopilotReadable } from "./use-make-copilot-readable";
+export { useMakeCopilotDocumentReadable } from "./use-make-copilot-document-readable";