+++ CopilotKit/packages/react-core/src/hooks/use-copilot-chat.ts
@@ -2,6 +2,7 @@ import { useMemo, useContext } from "react";
 import {
   CopilotContext,
   CopilotContextParams,
+  copilotApiConfigExtrapolator,
 } from "../context/copilot-context";
 import { useChat } from "ai/react";
 import { ChatRequestOptions, CreateMessage, Message } from "ai";
@@ -34,6 +35,7 @@ export function useCopilotChat({
     getContextString,
     getChatCompletionFunctionDescriptions,
     getFunctionCallHandler,
+    copilotApiConfig,
   } = useContext(CopilotContext);
 
   const systemMessage: Message = useMemo(() => {
@@ -57,13 +59,13 @@ export function useCopilotChat({
 
   const { messages, append, reload, stop, isLoading, input, setInput } =
     useChat({
+      api: copilotApiConfigExtrapolator(copilotApiConfig).chatApiEndpoint,
       id: options.id,
       initialMessages: initialMessagesWithContext,
       experimental_onFunctionCall: getFunctionCallHandler(),
       body: {
         id: options.id,
-        previewToken,
-        copilotkit_manually_passed_function_descriptions: functionDescriptions,
+        functions: functionDescriptions,
       },
     });
 
@@ -81,8 +83,6 @@ export function useCopilotChat({
     setInput,
   };
 }
-
-const previewToken = "TODO123";
 
 export function defaultSystemMessage(contextString: string): string {
   return `