+++ CopilotKit/packages/react-textarea/src/types/autosuggestions-config/suggestions-api-config.tsx
@@ -1,9 +1,7 @@
 import { MinimalChatGPTMessage } from "./subtypes/minimal-chat-gpt-message";
 import { MakeSystemPrompt } from "./subtypes/make-system-prompt";
-import { ChatlikeApiEndpoint } from "./subtypes/chatlike-api-endpoint";
 
 export interface SuggestionsApiConfig {
-  apiEndpoint: ChatlikeApiEndpoint;
   makeSystemPrompt: MakeSystemPrompt;
   fewShotMessages: MinimalChatGPTMessage[];
   forwardedParams: { [key: string]: any } | undefined;
@@ -46,7 +44,6 @@ export const defaultSuggestionsFewShotMessages: MinimalChatGPTMessage[] = [
   },
   {
     role: "user",
-    name: "TextBeforeCursor",
     content: "This morning I woke up and went straight to the grocery store.",
   },
   {
@@ -74,9 +71,6 @@ export const defaultSuggestionsFewShotMessages: MinimalChatGPTMessage[] = [
 ];
 
 export const defaultSuggestionsApiConfig: SuggestionsApiConfig = {
-  apiEndpoint: ChatlikeApiEndpoint.standardOpenAIEndpoint(
-    "/api/copilotkit_chatlike"
-  ),
   makeSystemPrompt: defaultSuggestionsMakeSystemPrompt,
   fewShotMessages: defaultSuggestionsFewShotMessages,
   forwardedParams: undefined,