+++ CopilotKit/packages/react-textarea/src/types/base/base-autosuggestions-config.tsx
@@ -2,6 +2,7 @@ import { BaseCopilotTextareaApiConfig } from "./autosuggestions-bare-function";
 
 export interface BaseAutosuggestionsConfig {
   textareaPurpose: string;
+  contextCategories: string[];
   debounceTime: number;
   acceptAutosuggestionKey: string;
   disableWhenEmpty: boolean;
@@ -15,6 +16,7 @@ export const defaultBaseAutosuggestionsConfig: Omit<
   "textareaPurpose" | "apiConfig"
 > = {
   debounceTime: 250,
+  contextCategories: ["global"],
   acceptAutosuggestionKey: "Tab",
   disableWhenEmpty: true,
   disabled: false,