+++ CopilotKit/examples/next-openai/CHANGELOG.md
@@ -1,5 +1,96 @@
 # web
 
+## 1.4.0-alpha.11
+
+### Minor Changes
+
+- copilot textarea documents - provide with code skeleton
+
+### Patch Changes
+
+- Updated dependencies
+  - @copilotkit/react-textarea@0.19.0-alpha.9
+  - @copilotkit/react-core@0.8.0-alpha.6
+  - @copilotkit/react-ui@0.9.0-alpha.11
+
+## 1.4.0-alpha.10
+
+### Minor Changes
+
+- minor bugfix
+
+### Patch Changes
+
+- Updated dependencies
+  - @copilotkit/react-textarea@0.19.0-alpha.8
+  - @copilotkit/react-core@0.8.0-alpha.5
+  - @copilotkit/react-ui@0.9.0-alpha.10
+
+## 1.4.0-alpha.9
+
+### Minor Changes
+
+- api endpoint centralization
+
+### Patch Changes
+
+- Updated dependencies
+  - @copilotkit/react-textarea@0.19.0-alpha.7
+  - @copilotkit/react-core@0.8.0-alpha.4
+  - @copilotkit/react-ui@0.9.0-alpha.9
+
+## 1.4.0-alpha.8
+
+### Minor Changes
+
+- package json
+
+### Patch Changes
+
+- Updated dependencies
+  - @copilotkit/react-textarea@0.19.0-alpha.6
+  - @copilotkit/react-ui@0.9.0-alpha.8
+  - @copilotkit/react-core@0.8.0-alpha.3
+
+## 1.4.0-alpha.7
+
+### Minor Changes
+
+- react-ui missing declaration
+
+### Patch Changes
+
+- Updated dependencies
+  - @copilotkit/react-textarea@0.19.0-alpha.5
+  - @copilotkit/react-ui@0.9.0-alpha.7
+  - @copilotkit/react-core@0.8.0-alpha.3
+
+## 1.4.0-alpha.6
+
+### Minor Changes
+
+- beta bump
+
+### Patch Changes
+
+- Updated dependencies
+  - @copilotkit/react-textarea@0.19.0-alpha.4
+  - @copilotkit/react-ui@0.9.0-alpha.6
+
+## 1.4.0-alpha.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @copilotkit/react-ui@0.9.0-alpha.5
+
+## 1.4.0-alpha.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @copilotkit/react-ui@0.9.0-alpha.4
+
 ## 1.4.0-alpha.3
 
 ### Patch Changes